feat: add rapid runner minigame

+++ src/lib/src/main/kotlin/work/lclpnet/ap2/ext/StatsExtensions.kt
@@ -10,7 +10,7 @@ import work.lclpnet.ap2.api.stats.CommonStats.Kills
 import work.lclpnet.ap2.api.stats.FFAStatsManager
 import work.lclpnet.ap2.ext.mc.playNotifySound
 import work.lclpnet.ap2.game.base.MapGameInstance
-import work.lclpnet.ap2.impl.game.data.type.PlayerRef
+import work.lclpnet.ap2.game.data.type.PlayerRef
 import work.lclpnet.kibu.hook.player.PlayerMoveCallback
 import work.lclpnet.kibu.hook.util.PositionRotation
 import work.lclpnet.kibu.translate.Translations
@@ -52,7 +52,9 @@ fun MapGameInstance.updateDistanceMoved(
     val dx = to.x() - from.x()
     val dz = to.z() - from.z()
 
-    stats.modify(player, CommonStats.DistanceMoved) { it + sqrt(dx * dx + dz * dz).toFloat() }
+    stats.modify(player, CommonStats.DistanceMoved) {
+        it + sqrt(dx * dx + dz * dz)
+    }
 }
 
 fun MapGameInstance.gainKill(player: ServerPlayer, stats: BaseStatsManager<ServerPlayer, PlayerRef>) {
